Heuristiques + mapa afegits

+++ src/algoritmes/aStar.java
@@ -9,7 +9,7 @@ public class AStar extends AlgoritmeCami {
     @Override
     public List<Casella> trobarCami(Casella inici, Casella fi, Mapa mapa, Heuristica heuristica) {
 
-        double hInicial = heuristica.calcular(inici, fi);
+        float hInicial = heuristica.calcular(inici, fi);
 
         // TODO: Implementación A*
