+++ src/heuristiques/Heuristica.java
@@ -3,5 +3,7 @@ package heuristiques;
 import model.Casella;
 
 public abstract class Heuristica {
-    public abstract double calcular(Casella actual, Casella objectiu);
+
+    public abstract float calcular(Casella actual, Casella objectiu);
+    
 }
