+++ src/heuristiques/HeuristicaManhattan.java
@@ -5,7 +5,7 @@ import model.Casella;
 public class HeuristicaManhattan extends Heuristica {
 
     @Override
-    public double calcular(Casella actual, Casella objectiu) {
+    public float calcular(Casella actual, Casella objectiu) {
         return Math.abs(actual.getX() - objectiu.getX()) +
                Math.abs(actual.getY() - objectiu.getY());
     }
